MyMangaTagger/requirements.txt
# MyMangaTagger requirements
# Python 3.12+

# Core dependencies (pinned after verified working state)
pillow==12.1.1
pyperclip==1.11.0
requests==2.33.0
tkinterdnd2==0.4.3
ttkbootstrap==1.20.2
